parser/requirements.txt
feedparser==6.0.10
psycopg2-binary==2.9.9
requests==2.31.0
python-dateutil==2.8.2